+++ src/lib/AsyncTypeaheadInput.tsx
@@ -152,7 +152,6 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadProps<T
             style={style}
             emptyLabel={emptyLabel}
             placeholder={placeholder}
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
             renderMenu={useGroupBy ? (results, menuProps) => renderMenu(results as LabelValueOption[], menuProps) : undefined}
           />
         </FormGroupLayout>