+++ src/lib/StaticTypeaheadInput.tsx
@@ -149,7 +149,6 @@ const StaticTypeaheadInput = <T extends FieldValues>(props: StaticTypeaheadInput
             emptyLabel={emptyLabel}
             placeholder={placeholder}
             {...reactBootstrapTypeaheadProps}
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
             renderMenu={useGroupBy ? (results, menuProps) => renderMenu(results as LabelValueOption[], menuProps) : undefined}
           />
         </FormGroupLayout>