+++ src/lib/helpers/typeahead.tsx
@@ -15,9 +15,14 @@ const convertTypeaheadOptionsToStringArray = (options: Option[]): string[] => {
 };
 
 const renderMenu = (results: LabelValueOption[], menuProps: RenderMenuProps): JSX.Element => {
+  const anonymousOptions: LabelValueOption[] = [];
   const groupedOptions = results.reduce((result, option) => {
     const groupKey = option.group?.name;
-    if (groupKey) (result[groupKey] = result[groupKey] || []).push(option);
+    if (groupKey) {
+      (result[groupKey] = result[groupKey] || []).push(option);
+    } else {
+      anonymousOptions.push(option);
+    }
     return result;
   }, {} as Record<string, LabelValueOption[]>);
   const groups = Object.keys(groupedOptions);
@@ -41,6 +46,12 @@ const renderMenu = (results: LabelValueOption[], menuProps: RenderMenuProps): JS
           ))}
           {index < groups.length - 1 && <Menu.Divider />}
         </React.Fragment>
+      ))}
+      {anonymousOptions.length > 0 && <Menu.Divider />}
+      {anonymousOptions.map((option) => (
+        <MenuItem key={option.value} option={option} position={position++} disabled={option.disabled}>
+          {option.label}
+        </MenuItem>
       ))}
     </Menu>
   );