+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - The `helpers/typeahead` format from `ts` to `tsx`
+- Corrected the cypress test running script
 
 ## [2.11.0] - 2024-10-04